ai-services/chatbot/requirements.txt
Flask==2.3.3
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
langchain==0.0.350
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
gunicorn==21.2.0
flask-cors==4.0.0
celery==5.3.4
